Revert "Fix Dual_Flange wet/dry to work like every other plugin."
This reverts commit 1427f53a058fe50f71343868a763af8fd7adb030.
It appears that the wet/dry is supposed to work differently for Dual Flange.
Probably the slider should be labeled as 'Mix' or something other than wet/dry
since the dry with the previous implementation will not always be 100% dry.
Gonna revert back to legacy behaviour since the fix will alter the bank presets
and does not allow for some precision settings.